requirements.txt
fastmcp
mcp
requests
langgraph
langchain-aws
langchain-core
boto3
python-dotenv

# FastAPI integration dependencies
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
